Use URNs throughout the app
Using the URNs inside the FileSystemService makes working with
paths a bit easier.

The frontend needs more generalised components. For example, the
<Toolbar> component is growing too much. It needs to accept
children as a prop instead of containing the logic for the actions.

Maybe SSE & reducers could be an interesting avenue to look into.
The client would push actions through trpc, but it wouldn't contain
any useful response beyond success/error. Then the event would get
processed by the server (e.g. a file gets moved) and once that's done,
an event is sent to the browser with SSE that would update the UI.